Merged PR 113: Fix for null reference exception being thrown by HttpContext
My fix (the one mentioned in the title) can be found in UserAccessor.cs and Startup.cs. The rest was agreed with Rafał and Filip and is not connected to the main issue. I removed "async void" from Blogger.cs, as it's not recommended. In EmailConfig, EmailService and appsettings.json you can find Filip's final (presumably) solution to the problem with sending mails from CRON.